add db init functions to backends - useful for library users
